fix(pools): move PoolListWithInitialData to shared lib, make poolListSlot optional
- Move PoolListWithInitialData from app-local _components to packages/lib/modules/pool/PoolList
- Make poolListSlot optional in PoolsPage with default Suspense+PoolList fallback
- Both frontend-v3 and beets-frontend-v3 now use the shared component
- Beets app also gets SSR pool data with Suspense streaming
